fix: usar BigDecimal para precio y descuento

+++ src/main/java/com/ggstore/ms_catalogo/dto/JuegoDTO.java
@@ -1,6 +1,7 @@
 package com.ggstore.ms_catalogo.dto;
 
 import lombok.Data;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -8,12 +9,12 @@ public class JuegoDTO {
     private UUID id;
     private String titulo;
     private String descripcion;
-    private Double precio;
-    private Double precioFinal;
+    private BigDecimal precio;
+    private BigDecimal precioFinal;
     private Integer stock;
     private String plataforma;
     private String imagenUrl;
-    private Integer descuentoPorcentaje;
+    private BigDecimal descuentoPorcentaje;
     private UUID categoriaId;
     private String categoriaNombre;
 }
